GameEditorView now functional.
GameObject set_world_position() and set_local_position() working as desired. Created a Timer class which can be used to track duration of different components of the engine.

GameObject add_child() is now a protected function. set_parent() now takes on the roll of setting a gameobjects parent, and adding that gameobject to the parents children set.